구르기 추가 및 몬스터 종류 추가

+++ SlashServer/SlashServer/GameObject.h
@@ -21,7 +21,7 @@ public:
 	XMFLOAT4X4 world_;
 	BoundingOrientedBox xmOOBB_;
 	BoundingOrientedBox xmOOBBTransformed_;
-	float lookDegree_;
+	double lookDegree_;
 	bool isActive_;
 	BYTE state_;
 	BYTE objectType_;
